Created Installer.
Added the ability to choose the SQL server address and login credentials inside Energize and Configurator. Still need to add functionality for the Server Console as well and add them as settings.
Removed the Excel Room Loader from the tabs page and added it to the tools drop down option and added the code and functionality behind that.

Added the start of whats going to be the State manager.